fix(iperf): fix iperf result error due to nano newlib and uint32_t data len

+++ examples/wifi/iperf/components/iperf/iperf.c
@@ -18,7 +18,7 @@
 typedef struct {
     iperf_cfg_t cfg;
     bool finish;
-    uint32_t total_len;
+    uint64_t total_len;
     uint32_t buffer_len;
     uint8_t *buffer;
     uint32_t sockfd;
